Added Balance changes

+++ Munching Monsters/Assets/Scripts/SaveLoad.cs
@@ -10,17 +10,17 @@ public class SaveLoad : MonoBehaviour
        // REMEMBER TO REMOVE
        PlayerPrefs.DeleteAll();
 
-       gVar.cookieSize = PlayerPrefs.GetInt("cookieSize", 1); 
-       gVar.cookieFlavours = PlayerPrefs.GetInt("cookieFlavours", 1); 
-       gVar.robotArmy = PlayerPrefs.GetInt("robotArmy", 0); 
+       gVar.cookieSize = PlayerPrefs.GetFloat("cookieSize", 1); 
+       gVar.cookieFlavours = PlayerPrefs.GetFloat("cookieFlavours", 1); 
+       gVar.robotArmy = PlayerPrefs.GetFloat("robotArmy", 0); 
        gVar.robotLvl = PlayerPrefs.GetInt("robotLvl", 1); 
        gVar.monsterLvl = PlayerPrefs.GetInt("monsterLvl", 1); 
-       gVar.gold = PlayerPrefs.GetInt("gold", 0);
+       gVar.gold = PlayerPrefs.GetFloat("gold", 0);
        gVar.monstersSpawned = PlayerPrefs.GetInt("monstersSpawned", 0);
-       gVar.cookieCost = PlayerPrefs.GetInt("cookieCost", 10 * gVar.cookieSize);
-       gVar.flavourCost = PlayerPrefs.GetInt("flavourCost", 20 * gVar.cookieFlavours);
-       gVar.robotCost = PlayerPrefs.GetInt("robotCost", 10 * gVar.robotArmy);
-       gVar.upgradeCost = PlayerPrefs.GetInt("upgradeCost", 20 * gVar.robotLvl);
+       gVar.cookieCost = PlayerPrefs.GetFloat("cookieCost", 10 * gVar.cookieSize);
+       gVar.flavourCost = PlayerPrefs.GetFloat("flavourCost", 20 * gVar.cookieFlavours);
+       gVar.robotCost = PlayerPrefs.GetFloat("robotCost", 10 * gVar.robotArmy);
+       gVar.upgradeCost = PlayerPrefs.GetFloat("upgradeCost", 20 * gVar.robotLvl);
     }
 
     // Update is called once per frame
